fix terra wallet connection
